Move the styling system up the stack to live right below Avalonia.Animation since it has no dependencies on anything further down the dependency chain
