+++ src/Avalonia.Base/AvaloniaObject.cs
@@ -703,7 +703,7 @@ namespace Avalonia
         /// <returns>The default value.</returns>
         private object GetDefaultValue(AvaloniaProperty property)
         {
-            if (property.Inherits && _inheritanceParent is AvaloniaObject aobj)
+            if (property.Inherits && InheritanceParent is AvaloniaObject aobj)
                 return aobj.GetValueInternal(property);
             return ((IStyledPropertyAccessor) property).GetDefaultValue(GetType());
         }
