+++ src/Avalonia.Controls/HotkeyManager.cs
@@ -55,13 +55,13 @@ namespace Avalonia.Controls
             public void Init()
             {
                 _hotkeySub = _control.GetObservable(HotKeyProperty).Subscribe(OnHotkeyChanged);
-                _parentSub = AncestorFinder.Create(_control, typeof (TopLevel)).Subscribe(OnParentChanged);
+                _parentSub = AncestorFinder.Create<TopLevel>(_control).Subscribe(OnParentChanged);
             }
 
-            private void OnParentChanged(IControl control)
+            private void OnParentChanged(TopLevel control)
             {
                 Unregister();
-                _root = (TopLevel) control;
+                _root = control;
                 Register();
             }
