+++ src/Avalonia.Controls/IControl.cs
@@ -16,37 +16,11 @@ namespace Avalonia.Controls
     /// </summary>
     public interface IControl : IVisual,
         IDataTemplateHost,
-        ILogical,
         ILayoutable,
         IInputElement,
         INamed,
-        IResourceNode,
-        IStyleable,
-        IStyleHost
+        IStyledElement
     {
-        /// <summary>
-        /// Occurs when the control has finished initialization.
-        /// </summary>
-        event EventHandler Initialized;
-
-        /// <summary>
-        /// Gets or sets the control's styling classes.
-        /// </summary>
-        new Classes Classes { get; set; }
-
-        /// <summary>
-        /// Gets or sets the control's data context.
-        /// </summary>
-        object DataContext { get; set; }
-
-        /// <summary>
-        /// Gets a value that indicates whether the element has finished initialization.
-        /// </summary>
-        bool IsInitialized { get; }
-
-        /// <summary>
-        /// Gets the control's logical parent.
-        /// </summary>
-        IControl Parent { get; }
+        new IControl Parent { get; }
     }
 }
