+++ src/Avalonia.Controls/Utils/AncestorFinder.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reactive;
 using System.Reactive.Disposables;
+using System.Reactive.Linq;
 using System.Reactive.Subjects;
 using System.Reflection;
 using System.Text;
@@ -14,15 +15,15 @@ namespace Avalonia.Controls.Utils
     {
         class FinderNode : IDisposable
         {
-            private readonly IControl _control;
+            private readonly IStyledElement _control;
             private readonly TypeInfo _ancestorType;
-            public IObservable<IControl> Observable => _subject;
-            private readonly Subject<IControl> _subject = new Subject<IControl>();
+            public IObservable<IStyledElement> Observable => _subject;
+            private readonly Subject<IStyledElement> _subject = new Subject<IStyledElement>();
 
             private FinderNode _child;
             private IDisposable _disposable;
 
-            public FinderNode(IControl control, TypeInfo ancestorType)
+            public FinderNode(IStyledElement control, TypeInfo ancestorType)
             {
                 _control = control;
                 _ancestorType = ancestorType;
@@ -33,7 +34,7 @@ namespace Avalonia.Controls.Utils
                 _disposable = _control.GetObservable(Control.ParentProperty).Subscribe(OnValueChanged);
             }
 
-            private void OnValueChanged(IControl next)
+            private void OnValueChanged(IStyledElement next)
             {
                 if (next == null || _ancestorType.IsAssignableFrom(next.GetType().GetTypeInfo()))
                     _subject.OnNext(next);
@@ -46,7 +47,7 @@ namespace Avalonia.Controls.Utils
                 }
             }
 
-            private void OnChildValueChanged(IControl control) => _subject.OnNext(control);
+            private void OnChildValueChanged(IStyledElement control) => _subject.OnNext(control);
 
 
             public void Dispose()
@@ -55,10 +56,15 @@ namespace Avalonia.Controls.Utils
             }
         }
 
-
-        public static IObservable<IControl> Create(IControl control, Type ancestorType)
+        public static IObservable<T> Create<T>(IStyledElement control)
+            where T : IStyledElement
         {
-            return new AnonymousObservable<IControl>(observer =>
+            return Create(control, typeof(T)).Cast<T>();
+        }
+
+        public static IObservable<IStyledElement> Create(IStyledElement control, Type ancestorType)
+        {
+            return new AnonymousObservable<IStyledElement>(observer =>
             {
                 var finder = new FinderNode(control, ancestorType.GetTypeInfo());
                 var subscription = finder.Observable.Subscribe(observer);
@@ -70,8 +76,6 @@ namespace Avalonia.Controls.Utils
                     finder.Dispose();
                 });
             });
-
-
         }
     }
 }
