+++ src/Avalonia.Styling/Controls/IPseudoClasses.cs
