+++ src/Avalonia.Styling/INamed.cs
