+++ src/Avalonia.Styling/Styling/IRequiresTemplateInStyle.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Avalonia.Styling
+{
+    /// <summary>
+    /// This is an interface for advanced scenarios to assist users in correct style development.
+    /// You as a user will not need to use this interface directly.
+    /// </summary>
+    public interface IRequiresTemplateInStyle
+    {
+    }
+}
