+++ src/Avalonia.Visuals/Avalonia.Visuals.csproj
@@ -5,6 +5,7 @@
   <ItemGroup> 
     <ProjectReference Include="..\Avalonia.Animation\Avalonia.Animation.csproj" />
     <ProjectReference Include="..\Avalonia.Base\Avalonia.Base.csproj" />
-  </ItemGroup>  
+    <ProjectReference Include="..\Avalonia.Styling\Avalonia.Styling.csproj" />
+  </ItemGroup>
   <Import Project="..\..\build\Rx.props" />
 </Project>
