+++ src/Avalonia.Visuals/Visual.cs
@@ -24,7 +24,7 @@ namespace Avalonia
     /// <see cref="IRenderer"/> to render the control. To traverse the visual tree, use the
     /// extension methods defined in <see cref="VisualExtensions"/>.
     /// </remarks>
-    public class Visual : Animatable, IVisual
+    public class Visual : StyledElement, IVisual
     {
         /// <summary>
         /// Defines the <see cref="Bounds"/> property.
